requirements.txt
discord.py
python-dotenv
PyNaCl
Pillow
python-slugify
requests
edge-tts
elevenlabs
openai